CMake/CI: Configure compiler warnings
This system for compiler warnings was lifted from Jason Turner's C++ starter
project.  It has two parts: a CMake function to configure all the warnings
for a target, and a fake "library" target that can be "linked" to other
targets so they can inherit its warnings.

I also added -DWARNINGS_AS_ERRORS=ON to the CI workflow to ensure that
they're always enforced by CI.